Standardized some naming and moved a few files around so that the server.js file would start without error
