+++ config/connections.js
@@ -3,14 +3,9 @@ require("dotenv").config();
 
 const sequelize = process.env.DB_URL
   ? new Sequelize(process.env.DB_URL)
-  : new Sequelize(
-      process.env.DB_NAME,
-      process.env.DB_USER,
-      process.env.DB_PASSWORD,
-      {
-        host: "localhost",
-        dialect: "postgres",
-      }
-    );
+  : new Sequelize(process.env.DB_NAME, process.env.DB_USER, process.env.DB_PASSWORD, {
+      host: "localhost",
+      dialect: "postgres",
+    });
 
 module.exports = sequelize;
