--- controllers/user-routes.js
@@ -1,51 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const { User } = require('../models');
-const bcrypt = require('bcrypt');
-
-router.post('/', async (req, res) => {
-    try {
-        const userData = await User.create(req.body);
-        req.session.save(() => {
-            req.session.user_id = userData.id;
-            req.session.logged_in = true;
-            res.status(200).json(userData);
-        });
-    } catch (err) {
-        res.status(400).json(err);
-    }
-});
-
-router.post('/login', async (req, res) => {
-    try {
-        const userData = await User.findOne({ where: { username: req.body.email } });
-
-        if (!userData) {
-            res.status(400).json({ message: 'Incorrect username or password, please try again' });
-            return;
-        }
-
-        const validPassword = await bcrypt.compare(req.body.password, userData.password);
-
-        if (!validPassword) {
-            res.status(400).json({ message: 'Incorrect username or password, please try again' });
-            return;
-        }
-
-        req.session.save(() => {
-            req.session.user_id = userData.id;
-            req.session.logged_in = true;
-            res.status(200).json({ user: userData, message: 'You are now logged in!' });
-        });
-    } catch (err) {
-        res.status(400).json(err);
-    }
-});
-
-router.post('/logout', (req, res) => {
-    req.session.destroy(() => {
-        res.status(204).end();
-    });
-});
-
-module.exports = router;
