+++ models/Workout.js
@@ -15,7 +15,7 @@ Workout.init(
       type: DataTypes.INTEGER,
       allowNull: false,
       references: {
-        model: "Exercise",
+        model: "exercise",
         key: "id",
       },
     },
@@ -23,7 +23,7 @@ Workout.init(
       type: DataTypes.INTEGER,
       allowNull: false,
       references: {
-        model: "User",
+        model: "user",
         key: "id",
       },
     },
