+++ models/WorkoutExercises.js
@@ -1,54 +1,54 @@
-const { Model, DataTypes } = require('sequelize');
-const sequelize = require('../config/connection');
+const { Model, DataTypes } = require("sequelize");
+const sequelize = require("../config/connections");
 
 class WorkoutExercises extends Model {}
 
 WorkoutExercises.init(
-    {
-        id: {
-            type: DataTypes.INTEGER,
-            allowNull: false,
-            primaryKey: true,
-            autoIncrement: true,
-        },
-        workout_id: {
-            type: DataTypes.INTEGER,
-            refrences: {
-                model: "workout",
-                key: "id",
-            },
-        },
-        exercise_id: {
-            type:DataTypes.INTEGER,
-            references: {
-                model: "exercise",
-                key: "id" 
-            },
-        },
-        sets: {
-            type: DataTypes.INTEGER,
-            allowNull: true,            
-        },
-        reps: {
-            type: DataTypes.INTEGER,
-            allowNull: true,
-        },
-        duration: {
-            type: DataTypes.DECIMAL,
-            allowNull: true,
-        },
-        notes: {
-            type: DataTypes.STRING,
-            allowNull: true,
-        },
+  {
+    id: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+      primaryKey: true,
+      autoIncrement: true,
     },
-    {
-      sequelize,
-      timestamps: false,
-      freezeTableName: true,
-      underscored: true,
-      modelName: "workoutexercises",
-    }
+    workout_id: {
+      type: DataTypes.INTEGER,
+      refrences: {
+        model: "workout",
+        key: "id",
+      },
+    },
+    exercise_id: {
+      type: DataTypes.INTEGER,
+      references: {
+        model: "exercise",
+        key: "id",
+      },
+    },
+    sets: {
+      type: DataTypes.INTEGER,
+      allowNull: true,
+    },
+    reps: {
+      type: DataTypes.INTEGER,
+      allowNull: true,
+    },
+    duration: {
+      type: DataTypes.DECIMAL,
+      allowNull: true,
+    },
+    notes: {
+      type: DataTypes.STRING,
+      allowNull: true,
+    },
+  },
+  {
+    sequelize,
+    timestamps: false,
+    freezeTableName: true,
+    underscored: true,
+    modelName: "workoutexercises",
+  }
 );
 
 module.exports = WorkoutExercises;
