+++ package.json
@@ -6,7 +6,6 @@
     "express": "^4.19.2",
     "express-handlebars": "^8.0.1",
     "express-session": "^1.18.0",
-    "handlebars": "^4.7.8",
     "pg": "^8.12.0",
     "sequelize": "^6.37.3"
   }
