+++ views/layout/main.handelbars
@@ -1,1 +1,43 @@
-//
+<html lang="en">
+
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Tech Blog</title>
+    <link rel="stylesheet" href="/css/jass.css" />
+  </head>
+
+  <body>
+    <header class="display-flex justify-space-between align-center p-2">
+      <h1>
+        Tech Blog
+      </h1>
+      {{#if logged_in}}
+        <nav>          
+          <button class="btn" id="home">Home</button>
+          <button class="btn" id="logout">Logout</button>
+          <button class="btn" id="dashboard">Dashboard</button>
+        </nav>
+      {{else}}
+        <nav>
+          <button class="btn" id="home">Home</button>
+          <button class="btn" id="login">Login</button>
+        </nav>
+      {{/if}}
+    </header>
+    <main class="container container-fluid mt-5">
+      {{{body}}}
+    </main>
+    <footer>
+
+    </footer>
+    {{! if the user is logged in, provide queryselector to the logout button }}
+    {{#if logged_in}}
+      <script src="js/logOut.js"></script>
+    {{else}}
+      {{! if the user is logged out, provide queryselector to the login button }}
+      <script src="js/logIn.js"></script>
+    {{/if}}
+  </body>
+
+</html>
